docker/docling/requirements.txt
docling>=2.66.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Embeddings
sentence-transformers>=2.6.0
llama-index-core>=0.14.0
llama-index-embeddings-huggingface>=0.5.0
einops>=0.8.0
